Move extending rolie feed to own source file.
